ml-service/requirements.txt
openai==0.23.1
nltk==3.7
fastapi==0.85.1
uvicorn==0.19.0
Flask==2.2.5
python-dotenv==1.0.0
